# statistics: use goroutine pool to improve performance