--- a/statistics/BUILD.bazel
+++ b/statistics/BUILD.bazel
@@ -121,6 +121,7 @@ go_test(
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_pingcap_log//:log",
         "@com_github_stretchr_testify//require",
+        "@com_github_tiancaiamao_gp//:gp",
         "@org_uber_go_goleak//:goleak",
         "@org_uber_go_zap//:zap",
     ],