<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #46267

Problem Summary:

### What is changed and how it works?

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)

```
before

goos: darwin
goarch: arm64
pkg: github.com/pingcap/tidb/statistics
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size100
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size100-8              38046         31401 ns/op
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size1000
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size1000-8              6194        180800 ns/op
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size10000
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size10000-8              100      11043109 ns/op
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size100000
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size100000-8               1    1995169417 ns/op

after

goos: darwin
goarch: arm64
pkg: github.com/pingcap/tidb/statistics
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size100
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size100-8              38656         29574 ns/op
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size1000
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size1000-8              7790        172931 ns/op
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size10000
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size10000-8              100      10659916 ns/op
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size100000
BenchmarkMergeGlobalStatsTopNByConcurrencyWithHists/Size100000-8               1    1647615250 ns/op
```


- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```